Refine SKU generation in bulkDuplicate method of MainProductAPIController
- Updated SKU generation logic to cast the extracted numeric value to an integer before padding, ensuring correct formatting for product variations.
- This change enhances the reliability of SKU creation during the product duplication process.

+++ app/Http/Controllers/API/MainProductAPIController.php
@@ -419,7 +419,7 @@ class MainProductAPIController extends AppBaseController
                     ->get();
 
                 // SKU do produto principal (ex: PRD000006 -> 06) para o padrão das variações
-                $mainShortCode = str_pad(preg_replace('/\D/', '', $newCode), 2, '0', STR_PAD_LEFT);
+                $mainShortCode = str_pad((int) preg_replace('/\D/', '', $newCode), 2, '0', STR_PAD_LEFT);
                 $variationIndex = 1;
 
                 foreach ($products as $product) {
